changing default profile picture to be local (was really slow for pytests), and implementing user_stats_v1 with Aleesha

+++ src/other.py
@@ -45,7 +45,9 @@ def clear_v1():
 
     images = glob.glob('src/static/*')
     for pic in images:
-        os.remove(pic)
+        # Saves the default image for profile pictures.
+        if pic != 'src/static/default.jpg':
+            os.remove(pic)
 
     data_store.set(store)
